chore: 로또 생성 데이터명 변경
