+++ data/src/main/java/com/junjange/data/datasource/LotteryDataSource.kt
@@ -1,7 +1,7 @@
 package com.junjange.data.datasource
 
 import com.junjange.data.model.remote.LotteryGetEntity
-import com.junjange.data.model.remote.LotteryRandomNumbersEntity
+import com.junjange.data.model.remote.LotteryNumbersEntity
 
 interface LotteryDataSource {
 
@@ -16,6 +16,6 @@ interface LotteryDataSource {
         sixthNum: Int
     ): Result<Unit>
 
-    suspend fun getLotteryRandom(): Result<LotteryRandomNumbersEntity>
+    suspend fun getLotteryRandom(): Result<LotteryNumbersEntity>
 
 }
