+++ data/src/main/java/com/junjange/data/mapper/LotteryNumbersMapper.kt
@@ -1,17 +1,15 @@
 package com.junjange.data.mapper
 
-import com.junjange.data.model.remote.LotteryGetNumbersEntity
-import com.junjange.domain.model.LotteryGetNumbers
+import com.junjange.data.model.remote.LotteryNumbersEntity
+import com.junjange.domain.model.LotteryNumbers
 
-internal fun List<LotteryGetNumbersEntity>.toDomain() = map {
-    LotteryGetNumbers(
-        firstNum = it.firstNum,
-        secondNum = it.secondNum,
-        thirdNum = it.thirdNum,
-        fourthNum = it.fourthNum,
-        fifthNum = it.fifthNum,
-        sixthNum = it.sixthNum,
-        correctNumbers = it.correctNumbers,
-        rank = it.rank
-    )
-}
+fun LotteryNumbersEntity.toDomain() = LotteryNumbers(
+    round = round,
+    winningDate = winningDate,
+    firstNum = firstNum,
+    secondNum = secondNum,
+    thirdNum = thirdNum,
+    fourthNum = fourthNum,
+    fifthNum = fifthNum,
+    sixthNum = sixthNum,
+)
