+++ data/src/main/java/com/junjange/data/repository/LotteryRepositoryImpl.kt
@@ -3,7 +3,7 @@ package com.junjange.data.repository
 import com.junjange.data.datasource.LotteryDataSource
 import com.junjange.data.mapper.toDomain
 import com.junjange.domain.model.LotteryGet
-import com.junjange.domain.model.LotteryRandomNumbers
+import com.junjange.domain.model.LotteryNumbers
 import com.junjange.domain.repository.LotteryRepository
 import javax.inject.Inject
 
@@ -32,7 +32,7 @@ internal class LotteryRepositoryImpl @Inject constructor(
         sixthNum = sixthNum
     )
 
-    override suspend fun getLotteryRandom(): Result<LotteryRandomNumbers> =
+    override suspend fun getLotteryRandom(): Result<LotteryNumbers> =
         dataSource.getLotteryRandom().mapCatching { it.toDomain() }
 
 }
