+++ remote/src/main/java/com/junjange/remote/datasource/LotteryDataSourceImpl.kt
@@ -2,7 +2,7 @@ package com.junjange.remote.datasource
 
 import com.junjange.data.datasource.LotteryDataSource
 import com.junjange.data.model.remote.LotteryGetEntity
-import com.junjange.data.model.remote.LotteryRandomNumbersEntity
+import com.junjange.data.model.remote.LotteryNumbersEntity
 import com.junjange.remote.api.ApiService
 import com.junjange.remote.model.request.LotteryRandomRequest
 import com.junjange.remote.model.response.toData
@@ -38,7 +38,7 @@ internal class LotteryDataSourceImpl @Inject constructor(
         }
 
 
-    override suspend fun getLotteryRandom(): Result<LotteryRandomNumbersEntity> = runCatching {
+    override suspend fun getLotteryRandom(): Result<LotteryNumbersEntity> = runCatching {
         apiService.getLotteryRandom().data.toData()
     }
 }
